Our app is a Fortune Cookie inventory system. Yep, we've finally hit the big time: working for a company that can tell you your future, wrapped up inside a cheap cookie shell.
There are six different fortune categories that are loaded from the database. And if you click on any of these, we see all of the fortunes for the category and how many have been printed.

The project is a small Symfony app - but all the Doctrine stuff translates to any app using the Doctrine ORM. We have 2 entities: Category and FortuneCookie with a ManyToOne relation from FortuneCookie to the Category.

For our homepage, we're using the entity manager to fetch the Category's repository and call the built-in findAll function. This returns every Category, and so far, it lets us be lazy and avoid writing a custom query. The template loops over these and prints them out.