LEFT JOIN, check! And just like normal SQL, we can use the fc alias from the joined table to update the WHERE clause. I'll break this onto multiple lines for my personal sanity and then add OR fc.fortune LIKE :searchTerm because fortune is the name of the property on FortuneCookie that holds the message.
Moment of truth! We've got a match! Our fortunes are being searched with 'cat' term.

JOINing, but Querying for the same Data

Even though we now have a LEFT JOIN, the result of the query is no different: it still returns an array of Category objects. We can and will do some JOINs in the future that actually select data from the joined table. But if all you do is JOIN like we're doing here, it doesn't change the data that's returned.