In every query so far, Doctrine gives us objects. That's its default mode, but we can also easily use it to select specific fields.
On our category page, you can see how many of each fortune has been printed over time. At the top, let's total those numbers with a SUM() query and print it out.

In showCategoryAction(), create a new variable - dollar_fortunesPrinted, that'll be a number. And of course, we'll write a new query to get this. But instead of shoving this into CategoryRepository, this queries the FortuneCookie entity, so we'll use its repository instead. So, AppBundle:FortuneCookie, and we'll call a new countNumberPrintedForCategory method. Pass the category variable object as an argument.

This will return the raw, summed number. To actually use this, pass this into the template.
And now print it out: {{ fortunesPrinted }}, put that through Twig's number_format filter and add the word 'total'