I don't trust myself. So, var_dump() in the controller with our trusty die statement.
Refresh! Uh oh, that's an awesome error:

[Semantical Error] line 0, col 88 near 'cat.name FROM': Error 'cat'
is not defined.

Debugging Bad DQL Queries

This is what it looks like when you mess up your DQL. Doctrine does a really good job of lexing and parsing the DQL you give it, so when you make a mistake, it'll give you a pretty detailed error. Here, cat is not defined is because our query references cat with cat.name, but I haven't made any JOINs to create a cat alias. cat is not defined.

But real quick - go back to the error. If you scroll down the stack trace a little, you'll eventually see the full query:

SELECT SUM(fc.numberPrinted) as fortunesPrinted,
    AVG(fc.numberPrinted) fortunesAverage,
    cat.name
    FROM AppBundle\Entity\FortuneCookie fc
    WHERE fc.category = :category
For me, sometimes the top error is so small, it doesn't make sense. But if I look at it in context of the full query, it's a lot easier to figure out what mistake I made.