Making a Raw SQL Query
Back in FortuneCookieRepository, let's write some simple SQL to test with:

SELECT * FROM fortune_cookie;
When you use the DBAL, there are no entities and it doesn't know about any of our Doctrine annotations. Yep, we're talking to the raw tables and columns. So I used fortune_cookies because that's the name of the actual table in the database.