Next, we'll use the SQL to get a statement. So:
dollar_stmt = dollar_conn->prepare(dollar_sql);
And then we can execute() that, which runs the query but doesn't give you the result. To get that, call ->fetchAll() and var_dump() that.
Try it! And there it is: exactly what you'd expect with no effort at all. It's literally the results - in array format - from the raw SQL query. Doctrine isn't trying to hide this feature from you - just grab the Connection object and you're dangerous.