Enough with all this SQL stuff. Remember the query builder I was raving about earlier? I promised that one of its benefits is that, with a query builder, you can re-use parts of a query. But we don't have any of that right now.
Open up CategoryRepository. We have three methods, and all of them repeat the same leftJoin() to cat.fortuneCookies and the addSelect().
Ah, duplication! When you see duplication like this - whether it's a WHERE, an ORDER BY or a JOIN - there's a simple solution. Just add a new private function and have it add this stuff to the query builder.

Query-modifying Functions

Create a private function called addFortuneCookieJoinAndSelect(), because that's what it's going to do! It'll accept a QueryBuilder as an argument. Our goal is to, well, add the join to that. So I'll copy the 2 pieces that we want, add a , then paste it there. And just for convenience, let's return this too.

So, our function takes in a QueryBuilder, it modifies it, then it returns it so we can make any more changes. I'll be a nice programmer and add some PHPDoc.