Calling those Functions
The findAllOrdered() function is the one that fuels the homepage. So let's start here! Get rid of that duplicated leftJoin and addSelect. Instead, just call ->addFortuneCookieJoinAndSelect() and pass it the . So we create the query builder, do some things with it, but let our new function take care of the join stuff.

This should give us the exact same results. But you should never believe me, so let's go back and refresh the homepage. Yep, nice!