And one more time in findWithFortunesJoin(). Remove the duplication, create a  variable, return the last part of the query, and stick our magic line in the middle.
Try it! Refresh and click into a category. It all works. And you know, I feel a lot better. If there's one things I don't want to duplicate, it's query logic. I hope this looks really obvious to you - it's just a simple coding technique. But it's kind of amazing, because it's not something you can do easily with string queries. And it can really save you if once you've got complex WHERE clauses that need to be re-used. You don't want to screw that stuff up.