Enabling a Filter Globally
Through all of this, you might be asking: 'What good is a filter if I need to enable it all the time.' Well first, the nice thing about filters is that you do have this ability to enable or disable them if you need to.

To enable a filter globally, you just need to follow these same steps in the bootstrap of your app. To hook into the beginning process of Symfony, we'll need an event listener.

I did the hard-work already and created a class called BeforeRequestListener in sec/AppBundle/EventListener/BeforeRequestListener.php file.

For Symfony peeps, you'll recognize the code in my services.yml:

10 lines  app/config/services.yml
services:
    before_request_listener:
        class: AppBundle\EventListener\BeforeRequestListener
        arguments: ['dollardoctrine.orm.entity_manager']
        tags:
            -
                name: kernel.event_listener
                event: kernel.request
                method: onKernelRequest

It registers this as a service and the tags at the bottom says, 'Hey, when Symfony boots, like right at the very beginning, call the onKernelRequest method.' I'm also passing the EntityManager as the first argument to the __construct() function. Because, ya know, we need that to enable filters.

Let's go steal the enabling code from FortuneController, take it all out and paste it into onKernelRequest. Instead of simply dollarem, we have dollarthis->em, since it's set on a property

Let's try it! Even though we took the enable() code out of the controller, the numbers don't change: our filter is still working. If we click into 'Proverbs', we see only 1. But if I disable the filter, we see all 3.

That's it! You're dangerous. If you've ever built a multi-tenant site where almost every query has a filter, life just got easy.